Singleton RegistrationHandler
(1) Made the RegistrationHandler a singleton that is managed by AppFactory.

(2) Moved the register-new-user action into an _actions directory in
/app/user/registration/ so that it is co-located with where it is
used.

(3) Fixed the name of the registration-handler test. Previously it was
called register-new-user.test.ts, which is the wrong name.
