+++ src/app/user/registration/_actions/register-new-user.ts
@@ -1,0 +1,14 @@
+"use server";
+
+import APP from "@/lib/app";
+import {
+  RegistrationRequest,
+  RegistrationResponse,
+} from "@/lib/user/registration-handler";
+
+export async function registerNewUser(
+  request: RegistrationRequest,
+): Promise<RegistrationResponse> {
+  const handler = await APP.getRegistrationHandler();
+  return handler.handle(request);
+}
