+++ src/app/user/registration/_components/donor-form.tsx
@@ -14,7 +14,7 @@ import Link from "next/link";
 import { RoutePath } from "@/lib/route-path";
 import { useRouter } from "next/navigation";
 import { BarkFormOption } from "@/components/bark/bark-form";
-import { registerNewUser } from "@/lib/actions/register-new-user";
+import { registerNewUser } from "@/app/user/registration/_actions/register-new-user";
 import { RegistrationRequest } from "@/lib/user/registration-handler";
 import {
   DogAntigenPresence,
